auth/requirements.txt
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary
uvicorn
fastapi